Troubleshooting memory problems, init random forest

+++ src/features.py
@@ -6,7 +6,7 @@ import numpy as np
 def smiles_to_fp(smiles, radius=2, n_bits=2048):
     mol = Chem.MolFromSmiles(smiles)
     if mol is None:
-        return np.zeros(n_bits)  # handle invalid SMILES
+        return np.zeros(n_bits, dtype=np.uint8)  # handle invalid SMILES
     fp = AllChem.GetMorganFingerprintAsBitVect(mol, radius, nBits=n_bits)
-    return np.array(fp)
+    return np.array(fp, dtype=np.uint8)
